feat: split lavamoat init scripts
