+++ app/html/background/background.html
@@ -4,7 +4,9 @@
     <meta charset="utf-8" />
   </head>
   <body>
-    <script src="../../scripts/load/bootstrap.ts" defer></script>
+    <script src="../../scripts/load/init-statehooks.ts" defer></script>
+    <script src="../../scripts/load/init-sentry.ts" defer></script>
+    <script src="../../scripts/load/init-globals.ts" defer></script>
     <script src="../../scripts/load/background.ts" defer></script>
   </body>
 </html>
