+++ app/html/other/offscreen.html
@@ -5,6 +5,8 @@
     <title>MetaMask Offscreen Page</title>
   </head>
   <body>
+    <script src="../../scripts/load/init-statehooks.ts" defer></script>
+    <script src="../../scripts/load/init-globals.ts" defer></script>
     <script src="../../offscreen/offscreen.ts" defer></script>
   </body>
 </html>
