+++ app/html/partials/partial-body.html
@@ -14,5 +14,7 @@
   <div class="loading-timeout-message" aria-live="polite" role="status"></div>
 </div>
 <div id="popover-content"></div>
-<script src="../../scripts/load/bootstrap.ts" defer></script>
+<script src="../../scripts/load/init-statehooks.ts" defer></script>
+<script src="../../scripts/load/init-sentry.ts" defer></script>
+<script src="../../scripts/load/init-globals.ts" defer></script>
 <script src="../../scripts/load/ui.ts" defer></script>
