+++ app/scripts/load/init-statehooks.ts
@@ -1,0 +1,6 @@
+// WARNING: This code runs outside of LavaMoat.
+
+// The root compartment will populate this with hooks.
+global.stateHooks = global.stateHooks || ({} as typeof stateHooks);
+
+export {};
